[🔨Refactor] CustomPicker rename 및 파일 이동
